pipeLiner/requirements.txt
typer
rich
requests
pydantic
python-dotenv
fastapi
uvicorn[standard]
google.generativeai